feat: Convert mongo dependency to autoconfigure package
