+++ VSDS-LDESServer4J-infra-mongo/mongo-repository-library/pom.xml
@@ -3,13 +3,14 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>vsds-ldesserver4j</artifactId>
+        <artifactId>VSDS-LDESServer4J-infra-mongo</artifactId>
         <groupId>be.vlaanderen.informatievlaanderen</groupId>
         <version>1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>VSDS-LDESServer4J-application</artifactId>
+    <artifactId>mongo-repository-library</artifactId>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
@@ -18,27 +19,25 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-mongodb</artifactId>
+            <version>${spring-boot.version}</version>
+        </dependency>
+        <dependency>
             <groupId>be.vlaanderen.informatievlaanderen</groupId>
-            <artifactId>VSDS-LDESServer4J-interface-rest</artifactId>
+            <artifactId>VSDS-LDESServer4J-domain</artifactId>
             <version>${project.version}</version>
         </dependency>
-    </dependencies>
 
-    <build>
-        <finalName>${project.artifactId}</finalName>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>repackage</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+        </dependency>
+
+    </dependencies>
 
 </project>
