+++ VSDS-LDESServer4J-infra-mongo/mongo-repository-library/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/infra/mongo/config/EndpointConfig.java
@@ -1,0 +1,14 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.infra.mongo.config;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+import java.util.Properties;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class EndpointConfig extends Properties {
+    private final String endpoint;
+}
