+++ VSDS-LDESServer4J-infra-mongo/mongo-repository-spring-boot-autoconfigure/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/infra/mongo/EndpointProperties.java
@@ -1,0 +1,13 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.infra.mongo;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "ldes")
+@Getter
+@Setter
+public class EndpointProperties {
+
+    private String endpoint;
+}
