+++ VSDS-LDESServer4J-infra-mongo/mongo-repository-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+be.vlaanderen.informatievlaanderen.ldes.server.infra.mongo.MongoAutoConfiguration
