+++ VSDS-LDESServer4J-infra-mongo/pom.xml
@@ -8,6 +8,13 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>mongo-repository-library</module>
+        <module>mongo-repository-spring-boot-autoconfigure</module>
+        <module>mongo-repository-spring-boot-starter</module>
+    </modules>
 
     <artifactId>VSDS-LDESServer4J-infra-mongo</artifactId>
 
@@ -15,28 +22,5 @@
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
     </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-mongodb</artifactId>
-            <version>${spring-boot.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>be.vlaanderen.informatievlaanderen</groupId>
-            <artifactId>VSDS-LDESServer4J-domain</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-        </dependency>
-
-    </dependencies>
 
 </project>
